backend/requirements.txt
crewai
crewai[tools]
fastapi
uvicorn
python-dotenv
langchain-openai
langchain
langgraph
langsmith
mcp
pydantic
litellm
